update docs, organize and add new helpers for hashes
