+++ attributes/config.rb
@@ -1,4 +1,4 @@
-# config/logger configuration.
+# config and logger configuration.
 default['osquery']['options']['config_plugin']  = 'filesystem'
 default['osquery']['options']['logger_plugin']  = 'filesystem'
 
@@ -8,7 +8,3 @@ default['osquery']['options']['events_expiry']  = 3600
 default['osquery']['options']['verbose']        = false
 default['osquery']['options']['worker_threads'] = 2
 default['osquery']['options']['enable_monitor'] = false
-
-# extra options
-default['osquery']['syslog']['enabled'] = true
-default['osquery']['syslog']['filename'] = '/etc/rsyslog.d/60-osquery.conf'
