+++ attributes/file_paths.rb
@@ -1,13 +1,7 @@
 # FIM configuration.
 default['osquery']['fim_enabled'] = false
 default['osquery']['file_paths'] = {
-  homes: [
-    '/home/%/.ssh/%%'
-  ],
-  etc: [
-    '/etc/%%'
-  ],
-  tmp: [
-    '/tmp/%%'
-  ]
+  homes: ['/home/%/.ssh/%%'],
+  etc: ['/etc/%%'],
+  tmp: ['/tmp/%%']
 }
