+++ attributes/schedule.rb
@@ -1,4 +1,4 @@
-# Example schedule config.
+# example schedule config.
 default['osquery']['schedule'] = {
   info: {
     query: 'SELECT * FROM osquery_info;',
