+++ recipes/centos.rb
@@ -6,14 +6,12 @@
 #
 
 centos_version = node['platform_version'].split('.')[0].to_i
-repo_checksum = node['osquery']['repo']["el#{centos_version}_checksum"]
 repo_url = "#{osquery_s3}/centos#{centos_version}/noarch"
 centos_repo = "osquery-s3-centos#{centos_version}-repo-1-0.0.noarch.rpm"
-file_cache = Chef::Config['file_cache_path']
 
 remote_file "#{file_cache}/#{centos_repo}" do
   source "#{repo_url}/#{centos_repo}"
-  checksum repo_checksum
+  checksum repo_hashes[:centos][centos_version]
   notifies :install, 'rpm_package[osquery repo]', :immediately
   not_if { node['osquery']['repo']['internal'] }
   action :create
